+++ platform/ios/app/MBXContextSnapshotViewController.h
@@ -1,0 +1,17 @@
+//
+//  MBXContextSnapshotViewController.h
+//  iosapp
+//
+//  Created by Nadia Barbosa on 9/9/19.
+//  Copyright © 2019 Mapbox. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface MBXContextSnapshotViewController : UIViewController
+
+@end
+
+NS_ASSUME_NONNULL_END